dashboard.v1/back/requirements.txt
django==1.10.1
djangorestframework
djangorestframework-jwt